[Task 64993] applied feedback
