+++ dspace-api/src/main/java/org/dspace/statistics/SolrLoggerUsageEventListener.java
@@ -15,12 +15,10 @@ import org.dspace.eperson.EPerson;
 import org.dspace.services.model.Event;
 import org.dspace.statistics.service.SolrLoggerService;
 import org.dspace.usage.AbstractUsageEventListener;
-import org.dspace.usage.RestUsageSearchEvent;
 import org.dspace.usage.UsageEvent;
 import org.dspace.usage.UsageSearchEvent;
 import org.dspace.usage.UsageWorkflowEvent;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.util.CollectionUtils;
 
 /**
  * Simple SolrLoggerUsageEvent facade to separate Solr specific
@@ -57,28 +55,13 @@ public class SolrLoggerUsageEventListener extends AbstractUsageEventListener {
                                                    currentUser);
                     }
                 } else if (UsageEvent.Action.SEARCH == ue.getAction()) {
-                    if (ue instanceof UsageSearchEvent) {
-                        UsageSearchEvent usageSearchEvent = (UsageSearchEvent) ue;
-                        //Only log if the user has already filled in a query !
-                        if (!CollectionUtils.isEmpty(((UsageSearchEvent) ue).getQueries())) {
-                            solrLoggerService.postSearch(ue.getObject(), ue.getRequest(), currentUser,
-                                                         usageSearchEvent.getQueries(), usageSearchEvent.getRpp(),
-                                                         usageSearchEvent.getSortBy(),
-                                                         usageSearchEvent.getSortOrder(), usageSearchEvent.getPage(),
-                                                         usageSearchEvent.getScope());
-                        }
-                    } else {
-                        RestUsageSearchEvent restUsageSearchEvent = (RestUsageSearchEvent) ue;
-
-                        List<String> queries = new LinkedList<>();
-                        queries.add(restUsageSearchEvent.getQuery());
-                        solrLoggerService
-                            .postSearch(restUsageSearchEvent.getObject(), restUsageSearchEvent.getRequest(),
-                                        currentUser, queries, restUsageSearchEvent.getPage().getSize(),
-                                        restUsageSearchEvent.getSort().getBy(),
-                                        restUsageSearchEvent.getSort().getOrder(),
-                                        restUsageSearchEvent.getPage().getNumber(), restUsageSearchEvent.getScope());
-                    }
+                    UsageSearchEvent usageSearchEvent = (UsageSearchEvent) ue;
+                    List<String> queries = new LinkedList<>();
+                    queries.add(usageSearchEvent.getQuery());
+                    solrLoggerService.postSearch(usageSearchEvent.getObject(), usageSearchEvent.getRequest(),
+                                    currentUser, queries, usageSearchEvent.getPage().getSize(),
+                                    usageSearchEvent.getSort().getBy(), usageSearchEvent.getSort().getOrder(),
+                                    usageSearchEvent.getPage().getNumber(), usageSearchEvent.getScope());
                 } else if (UsageEvent.Action.WORKFLOW == ue.getAction()) {
                     UsageWorkflowEvent usageWorkflowEvent = (UsageWorkflowEvent) ue;
